new test spec has been added for unsupported file creator and xml and json file creators have been moved to folders
