+++ src/app/factory/json/json-file-creator.ts
@@ -1,6 +1,6 @@
-import { CsvFileModel } from "../models/csv-file-model";
-import { FileCreator } from "./file-creator";
-import { FileModel } from "./models/file-model";
+import { CsvFileModel } from "src/app/models/csv-file-model";
+import { FileCreator } from "../file-creator";
+import { FileModel } from "../models/file-model";
 
 export class JsonFileCreator extends FileCreator {
     extension: string = '.json'
