+++ src/app/factory/xml/xml-file-creator.ts
@@ -1,6 +1,6 @@
-import { CsvFileModel } from "../models/csv-file-model";
-import { FileCreator } from "./file-creator";
-import { FileModel } from "./models/file-model";
+import { CsvFileModel } from "../../models/csv-file-model";
+import { FileCreator } from "../file-creator";
+import { FileModel } from "../models/file-model";
 
 export class XmlFileCreator extends FileCreator {
     extension: string = '.xml'
